% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-static.R
\name{callStatic}
\alias{callStatic}
\title{Call a static method on a .NET type}
\usage{
callStatic(typename, methodName, ...)
}
\arguments{
\item{typename}{type name, possibly namespace and assembly qualified type name, e.g. 'My.Namespace.MyClass,MyAssemblyName'.}

\item{methodName}{the name of a static method of the type}

\item{...}{additional method arguments passed to .External (e.g., arguments to the method)}
}
\value{
an object resulting from the call. May be a `cobjRef` object, or a native R object for common types. Can be NULL.
}
\description{
Call a static method on a .NET type
}
\examples{
cTypename <- getRSharpSetting("testCasesTypeName")
callStatic(cTypename, "IsTrue", TRUE)
}
