% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-net-object.R
\name{castToRObject}
\alias{castToRObject}
\title{Create if possible an object of the R6 class `NetObject`}
\usage{
castToRObject(obj, recursive = TRUE)
}
\arguments{
\item{obj}{the presumed external pointer.}

\item{recursive}{logical; if TRUE, the function is applied recursively to the list elements.}
}
\value{
A `NetObject` R6 object if the argument is indeed an external pointer,
otherwise returned unchanged. If `recursive` is TRUE and `obj` is a list, the function is applied
recursively to the list elements.
}
\description{
Create if possible an object of the R6 class `NetObject`
}
\details{
Create if possible and adequate the R6 object of the class `NetObject`
that wraps the external pointer to a .NET object.
If `obj` is not a pointer, returns `obj` unchanged.
}
\examples{
castToRObject(1)
castToRObject("a")
castToRObject(TRUE)
castToRObject(FALSE)
castToRObject(1L)
castToRObject(1.1)
castToRObject(1.1 + 1i)
castToRObject(list(1, 2, 3))
castToRObject(data.frame(a = 1:3, b = c("a", "b", "c")))
}
