% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-net-object.R
\name{clrCall}
\alias{clrCall}
\title{Call a method on an object}
\usage{
clrCall(obj, methodName, ...)
}
\arguments{
\item{obj}{an object}

\item{methodName}{the name of a method of the object}

\item{...}{additional method arguments passed to .External}
}
\value{
an object resulting from the call. May be a CLR object, or a native R object for common types. Can be NULL.
}
\description{
Call a method on an object
}
\examples{
testClassName <- rSharpEnv$testObjectTypeName
testObj <- clrNew(testClassName)
clrCall(testObj, "GetFieldIntegerOne")
}
