% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSharp-exported.R, R/utilities-type.R
\name{clrGetConstructors}
\alias{clrGetConstructors}
\title{List the public constructors of a CLR Type}
\usage{
clrGetConstructors(type)

clrGetConstructors(type)
}
\arguments{
\item{type}{CLR Type, or a (character) type name that can be successfully parsed}
}
\value{
a list of constructor signatures

a list of constructor signatures
}
\description{
List the public constructors of a CLR Type
}
\examples{
\dontrun{
library(rSharp)
testClassName <- "ClrFacade.TestObject"
clrGetConstructors(testClassName)
}
\dontrun{
library(rClr)
testClassName <- "Rclr.TestObject"
clrGetConstructors(testClassName)
}
}
