% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSharp-exported.R, R/utilities-enum.R
\name{clrGetEnumNames}
\alias{clrGetEnumNames}
\title{Gets the names of a CLR Enum value type}
\usage{
clrGetEnumNames(enumType)

clrGetEnumNames(enumType)
}
\arguments{
\item{enumType}{a CLR object, System.Type or type name, possibly namespace and assembly qualified type name, e.g. 'My.Namespace.MyClass,MyAssemblyName'.}
}
\value{
a character vector of the names for the enum

a character vector of the names for the enum
}
\description{
Gets the names of a CLR Enum value type

Gets the names of a CLR Enum value type
}
\examples{
\dontrun{
library(rSharp)
clrGetEnumNames("ClrFacade.TestEnum")
}
\dontrun{
library(rClr)
clrGetEnumNames("Rclr.TestEnum")
}
}
