% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-assembly.R
\name{clrGetLoadedAssemblies}
\alias{clrGetLoadedAssemblies}
\title{List the names of loaded CLR assemblies}
\usage{
clrGetLoadedAssemblies(fullname = FALSE, filenames = FALSE)
}
\arguments{
\item{fullname}{Boolean. Should the full name of the assemblies be returned?
Default is `FALSE`}

\item{filenames}{Boolean. If `TRUE`, return a data frame where the second
column is the URI (usually file path) of the loaded assembly. Default is `FALSE`}
}
\value{
The names of loaded CLR assemblies
}
\description{
List the names of loaded CLR assemblies
}
\examples{
\dontrun{
library(rClr)
clrGetLoadedAssemblies()
}
}
