% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSharp-exported.R
\name{clrGetMethods}
\alias{clrGetMethods}
\title{List the instance methods of a .NET object}
\usage{
clrGetMethods(clrobj, contains = "")
}
\arguments{
\item{clrobj}{CLR object}

\item{contains}{a string that the methods names returned must contain}
}
\value{
a list of names of the methods of the .NET object
}
\description{
List the instance methods of a .NET object
}
\examples{
testClassName <- "ClrFacade.TestObject"
testObj <- clrNew(testClassName)
clrGetMethods(testObj)
clrGetMethods(testObj, "Get")
}
