% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSharp-exported.R
\name{clrGetNativeLibName}
\alias{clrGetNativeLibName}
\title{Gets the name of the native library currently loaded.}
\usage{
clrGetNativeLibName()
}
\value{
the name of the native library currently loaded: rSharpMs or rSharpUX
}
\description{
Gets the name of the native library currently loaded. Used only for unit tests.
}
