% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-net-object.R
\name{clrIs}
\alias{clrIs}
\title{Check whether an object is of a certain type}
\usage{
clrIs(obj, type)
}
\arguments{
\item{obj}{an object}

\item{type}{the object type to check for. It can be a character, of a object of CLR type System.RuntimeType}
}
\value{
TRUE or FALSE
}
\description{
This function is meant to match the behavior of the 'is' keyword in C#.
}
\examples{
testClassName <- "ClrFacade.TestObject"
testObj <- clrNew(testClassName)
clrIs(testObj, testClassName)
clrIs(testObj, "System.Object")
}
