% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-net-object.R
\name{clrSet}
\alias{clrSet}
\title{Sets the value of a field or property of an object or class}
\usage{
clrSet(objOrType, name, value)
}
\arguments{
\item{objOrType}{a CLR object, or type name, possibly namespace and assembly qualified type name, e.g. 'My.Namespace.MyClass,MyAssemblyName'.}

\item{name}{the name of a field/property of the object}

\item{value}{the value to set the field with}
}
\description{
Sets the value of a field or property of an object or class
}
\examples{
testClassName <- rSharpEnv$testObjectTypeName
testObj <- newObjectFromName(testClassName)
clrReflect(testObj)
clrSet(testObj, "FieldIntegerOne", as.integer(42))
clrSet(testClassName, "StaticPropertyIntegerOne", as.integer(42))
}
