% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSharp-exported.R, R/utilities-runtime.R
\name{clrShutdown}
\alias{clrShutdown}
\title{Shuts down the current runtime.}
\usage{
clrShutdown()

clrShutdown()
}
\value{
nothing is returned by this function

nothing is returned by this function
}
\description{
Shuts down the current runtime.
}
