% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-net-object.R
\name{clrTypename}
\alias{clrTypename}
\title{Gets the type name of an object}
\usage{
clrTypename(clrobj)
}
\arguments{
\item{clrobj}{.NET object}
}
\value{
type name
}
\description{
Gets the .NET type name of an object, given an S4 clrobj object
}
\examples{
testClassName <- "ClrFacade.TestObject"
testObj <- clrNew(testClassName)
clrTypename(testObj)
}
