% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-net-object.R
\name{createNetObject}
\alias{createNetObject}
\title{Create if possible an object of the R6 class `NetObject`}
\usage{
createNetObject(obj)
}
\arguments{
\item{obj}{the presumed external pointer.}
}
\value{
A `NetObject` R6 object if the argument is indeed an external pointer,
otherwise returned unchanged.
}
\description{
Create if possible an object of the R6 class `NetObject`
}
\details{
Create if possible and adequate the R6 object of the class `NetObject`
that wraps the external pointer to a .NET object.
If `obj` is not a pointer, returns `obj` unchanged.
}
