% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSharp-internal.R
\name{.clrTypeNameExtPtr}
\alias{.clrTypeNameExtPtr}
\title{Gets the type name of an object}
\usage{
.clrTypeNameExtPtr(extPtr)
}
\arguments{
\item{extPtr}{external pointer to a .NET object (not a cobjRef S4 or a `NetObject` object)}
}
\value{
a character string, the type name
}
\description{
Gets the type name of an object, given the SEXP external pointer to this .NET object.
}
\examples{
testClassName <- "ClrFacade.TestObject"
testObj <- newObjectFromName(testClassName)
.clrTypeNameExtPtr(testObj@clrobj)
}
