% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSharp-internal.R
\name{.extractPointersFromArgs}
\alias{.extractPointersFromArgs}
\title{Extract the pointers from the arguments}
\usage{
.extractPointersFromArgs(args)
}
\arguments{
\item{args}{List of arguments}
}
\description{
Extract the pointers from the arguments
}
\details{
Iterates through the list of arguments and replaces `NetObjects` by their
pointers. Required to pass to ClrFacade.
}
