% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSharp-internal.R
\name{.getCurrentConvertedObject}
\alias{.getCurrentConvertedObject}
\title{System function to get a direct access to an object}
\usage{
.getCurrentConvertedObject()
}
\value{
a `cobjRef` S4 object
}
\description{
This function is highly unlikely to be of any use to an end user, even an advanced one.
This is indirectly needed to unlock the benefits of using R.NET convert data structures between R and .NET.
}
