% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-type.R
\name{getConstructors}
\alias{getConstructors}
\title{List the public constructors of a CLR Type}
\usage{
getConstructors(type)
}
\arguments{
\item{type}{.NET Type, or a (character) type name that can be successfully parsed}
}
\value{
a list of constructor signatures
}
\description{
List the public constructors of a CLR Type
}
\examples{
testClassName <- "ClrFacade.TestObject"
getConstructors(testClassName)
}
