% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-enum.R
\name{getEnumNames}
\alias{getEnumNames}
\title{Gets the names of a .NET Enum value type}
\usage{
getEnumNames(enumType)
}
\arguments{
\item{enumType}{a .NET object, System.Type or type name, possibly namespace and assembly qualified type name, e.g. 'My.Namespace.MyClass,MyAssemblyName'.}
}
\value{
a character vector of the names for the enum
}
\description{
Gets the names of a .NET Enum value type
}
\examples{
getEnumNames("Rclr.TestEnum")
}
