% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSharp-exported.R
\name{getFields}
\alias{getFields}
\title{List the instance fields of a .NET object}
\usage{
getFields(clrobj, contains = "")
}
\arguments{
\item{clrobj}{CLR object}

\item{contains}{a string that the field names returned must contain}
}
\value{
a list of names of the fields of the .NET object
}
\description{
List the instance fields of a .NET object
}
\examples{
testClassName <- rSharpEnv$testObjectTypeName
testObj <- newObjectFromName(testClassName)
getFields(testObj)
getFields(testObj, "ieldInt")
}
