% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-assembly.R
\name{getLoadedAssemblies}
\alias{getLoadedAssemblies}
\title{List the names of loaded assemblies}
\usage{
getLoadedAssemblies(fullname = FALSE, filenames = FALSE)
}
\arguments{
\item{fullname}{should the full name of the assemblies be returned. `FALSE` by default.}

\item{filenames}{if TRUE, return a data frame where the second column is the URI (usually file path) of the loaded assembly. `FALSE` by default.}
}
\value{
the names of loaded assemblies
}
\description{
List the names of loaded assemblies
}
\examples{
getLoadedAssemblies()
}
