% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSharp-exported.R
\name{getNativeLibsPath}
\alias{getNativeLibsPath}
\title{Architecture dependent path to the rSharp native library}
\usage{
getNativeLibsPath(pkgName)
}
\arguments{
\item{pkgName}{the name of a package, e.g. 'rSharp'}
}
\value{
the prospective path in which a native library would be found, e.g. c:/R/library/rSharp/libs/x64
}
\description{
Guess the directory where to expect the architecture dependent native library of a specified package
e.g. for the package rSharp, ${R_HOME}/library/rSharp/libs/x64
This is a utility that is not specific to the rSharp package
}
