% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSharp-env.R
\name{getRSharpSetting}
\alias{getRSharpSetting}
\title{getRSharpSetting}
\usage{
getRSharpSetting(settingName)
}
\arguments{
\item{settingName}{String name of the setting}
}
\value{
Value of the setting stored in rSharpEnv. If the setting does not exist, an error is thrown.
}
\description{
Get the value of a global rSharp setting.
}
\examples{
getRSharpSetting("nativePkgName")
}
