% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSharp-exported.R, R/utilities-diagnostics.R
\name{getSexpType}
\alias{getSexpType}
\title{Get the type code for a SEXP}
\usage{
getSexpType(sexp)

getSexpType(sexp)
}
\arguments{
\item{sexp}{an R object}
}
\value{
the type code, an integer, as defined in Rinternals.h

the type code, an integer, as defined in Rinternals.h
}
\description{
Get the type code for a SEXP, as returned by the TYPEOF macro

Get the type code for a SEXP, as returned by the TYPEOF macro
}
