% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-static.R
\name{getStatic}
\alias{getStatic}
\title{Gets the value of a static field or property of a class}
\usage{
getStatic(type, name)
}
\arguments{
\item{type}{Type name, possibly namespace and assembly qualified type name, e.g. 'My.Namespace.MyClass,MyAssemblyName'.}

\item{name}{the name of a field/property  of the object}
}
\value{
An object resulting from the call. May be a `NetObject` object, or a native R object for common types. Can be NULL.
}
\description{
Gets the value of a static field or property of a class
}
\examples{
testClassName <- rSharpEnv$testObjectTypeName
getStatic(testClassName, "StaticPropertyIntegerOne")
}
