% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-static.R
\name{getStaticMembers}
\alias{getStaticMembers}
\title{Gets the static members for a type}
\usage{
getStaticMembers(objOrType)
}
\arguments{
\item{objOrType}{a .NET object, or type name, possibly namespace and assembly qualified type name, e.g. 'My.Namespace.MyClass,MyAssemblyName'.}
}
\description{
Gets the static members for a type
}
\examples{
cTypename <- getRSharpSetting("testCasesTypeName")
getStaticMembers(cTypename)
}
