% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-static.R
\name{getStaticMethods}
\alias{getStaticMethods}
\title{Gets the static methods for a type}
\usage{
getStaticMethods(objOrType, contains = "")
}
\arguments{
\item{objOrType}{a .NET object, or type name, possibly namespace and assembly qualified type name, e.g. 'My.Namespace.MyClass,MyAssemblyName'.}

\item{contains}{a string that the property names returned must contain}
}
\description{
Gets the static methods for a type
}
