% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSharp-exported.R
\name{getType}
\alias{getType}
\title{Gets the type of a .NET object given its type name}
\usage{
getType(objOrTypename)
}
\arguments{
\item{objOrTypename}{a character vector of length one. It can be the full file name of the assembly to load, or a fully qualified assembly name, or as a last resort a partial name.}
}
\value{
the .NET Type.
}
\description{
Gets the type of a .NET object given its type name
}
