% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-assembly.R
\name{isAssemblyLoaded}
\alias{isAssemblyLoaded}
\title{Is the assembly loaded?}
\usage{
isAssemblyLoaded(assemblyName)
}
\arguments{
\item{assemblyName}{The name of the assembly, e.g. 'ClrFacade'}
}
\value{
TRUE if the assembly is loaded, FALSE otherwise
}
\description{
Is the assembly loaded?
}
\examples{
isAssemblyLoaded("ClrFacade")
}
