% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-assembly.R
\name{loadAssembly}
\alias{loadAssembly}
\title{Loads a .NET assembly.}
\usage{
loadAssembly(name)
}
\arguments{
\item{name}{a character vector of length one. It can be the full file name of the assembly to load, or a fully qualified assembly name, or as a last resort a partial name.}
}
\value{
Name of the loaded assembly, if successfull.
}
\description{
Loads a .NET assembly.
}
\details{
Note that this is loaded in the single application domain that is created by rSharp, not a separate application domain.
}
\examples{
\dontrun{
f <- file.path("SomeDirectory", "YourDotNetBinaryFile.dll")
f <- path.expand(f)
stopifnot(file.exists(f))
loadAssembly(f)
}
}
\seealso{
\code{\link{.C}} which this function wraps
}
