% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-net-object.R
\name{newObjectFromName}
\alias{newObjectFromName}
\title{Create a new NetObject R6 object given the type name.}
\usage{
newObjectFromName(typename, ...)
}
\arguments{
\item{typename}{type name, possibly namespace and assembly qualified type name, e.g. 'My.Namespace.MyClass,MyAssemblyName'.}

\item{...}{additional method arguments passed to the object constructor via the call to .External}
}
\value{
a `NetObject` R6 object
}
\description{
Create a new NetObject R6 object given the type name.
}
\examples{
testClassName <- getRSharpSetting("testObjectTypeName")
testObj <- newObjectFromName(testClassName)
# object with a constructor that has parameters
testObj <- newObjectFromName(testClassName, as.integer(123))
}
