% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-net-object.R
\name{newPointerFromName}
\alias{newPointerFromName}
\title{Create a new external pointer to a .NET object given the type name.}
\usage{
newPointerFromName(typename, ...)
}
\arguments{
\item{typename}{type name, possibly namespace and assembly qualified type name, e.g. 'My.Namespace.MyClass,MyAssemblyName'.}

\item{...}{additional method arguments passed to the object constructor via the call to .External}
}
\value{
an external pointer to a .NET object
}
\description{
Create a new external pointer to a .NET object given the type name.
}
\examples{
testClassName <- getRSharpSetting("testObjectTypeName")
testPtr <- newPointerFromName(testClassName)
# Now we can create a NetObject from the pointer
testObj <- NetObject$new(testPtr)
}
