% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSharp-exported.R, R/utilities-diagnostics.R
\name{peekClrArgs}
\alias{peekClrArgs}
\title{Peek into the types of CLR objects arguments are converted to by rSharp}
\usage{
peekClrArgs(...)

peekClrArgs(...)
}
\arguments{
\item{...}{method arguments passed to .External}
}
\value{
a character message with type information about each argument.

a character message with type information about each argument.
}
\description{
Advanced use only, to diagnose unexpected conditions in CLR method calls. Most users would not ever need it.

Advanced use only, to diagnose unexpected conditions in CLR method calls. Most users would not ever need it.
}
\examples{
\dontrun{
library(rSharp)
peekClrArgs("a", numeric(0))
}
\dontrun{
library(rClr)
peekClrArgs("a", numeric(0))
}
}
