% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-diagnostics.R
\name{printTraceback}
\alias{printTraceback}
\title{Prints the last .NET exception}
\usage{
printTraceback()
}
\description{
This is roughly the equivalent of the traceback function of R.
}
\examples{
callStatic(getRSharpSetting("testCasesTypeName"), "ThrowException", 10L) # will be truncated by the Rf_error API
printTraceback() # prints the full stack trace
}
