% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSharp-env.R
\name{rSharpSettingNames}
\alias{rSharpSettingNames}
\title{Names of the settings stored in rSharpEnv Can be used with `getRSharpSetting()`}
\usage{
rSharpSettingNames
}
\description{
Names of the settings stored in rSharpEnv Can be used with `getRSharpSetting()`
}
