% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-diagnostics.R
\name{rToDotNetType}
\alias{rToDotNetType}
\title{Gets the type of a .NET object resulting from converting an R object}
\usage{
rToDotNetType(x)
}
\arguments{
\item{x}{An R object}
}
\value{
A list, with columns including mode, type, class, length and the string of the corresponding .NET type.
}
\description{
Gets the type of a .NET object resulting from converting an R object. This function is mostly for documentation purposes, but may be of use to end users.
}
\examples{
rToDotNetType(1)
}
