% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{toStringNET}
\alias{toStringNET}
\title{Calls the ToString method of an object}
\usage{
toStringNET(x)
}
\arguments{
\item{x}{any R object, which is converted to a .NET object on which to call ToString}
}
\value{
The string representation of the object in .NET
}
\description{
Calls the ToString method of an object as represented in the .NET.
This function is here to help quickly test object equivalence from the R interpreter,
 for instance on the tricky topic of date-time conversions
}
\examples{
library(rSharp)
dt <- as.POSIXct("2001-01-01 02:03:04", tz = "UTC")
toStringNET(dt)
}
