% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-assembly.R
\name{clrGetTypesInAssembly}
\alias{clrGetTypesInAssembly}
\title{Get a list of CLR type names exported by an assembly}
\usage{
clrGetTypesInAssembly(assemblyName)
}
\arguments{
\item{assemblyName}{the name of the assembly}
}
\value{
The names of the types exported by the assembly
}
\description{
Get a list of CLR type names exported by an assembly
}
\examples{
\dontrun{
library(rClr)
clrGetLoadedAssemblies()
clrGetTypesInAssembly("ClrFacade")
}
}
