% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSharp-exported.R
\name{clrNew}
\alias{clrNew}
\title{Create a new CLR object}
\usage{
clrNew(typename, ...)
}
\arguments{
\item{typename}{type name, possibly namespace and assembly qualified type name, e.g. 'My.Namespace.MyClass,MyAssemblyName'.}

\item{...}{additional method arguments passed to the object constructor via the call to .External}
}
\value{
a CLR object
}
\description{
Create a new CLR object
}
\examples{
testClassName <- getRSharpSetting("testObjectTypeName")
testObj <- newObjectFromName(testClassName)
# object with a constructor that has parameters
testObj <- clrNew(testClassName, as.integer(123))
}
