% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-runtime.R
\name{clrTraceback}
\alias{clrTraceback}
\title{Prints the last CLR exception}
\usage{
clrTraceback()
}
\description{
This is roughly the equivalent of the traceback function of R.
}
\examples{
\dontrun{
callStatic("Rclr.TestCases", "ThrowException", 10L) # will be truncated by the Rf_error API
clrTraceback() # prints the full stack trace
}
}
