% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSharp-exported.R, R/utilities-net-object.R
\name{clrTypename}
\alias{clrTypename}
\title{Gets the type name of an object}
\usage{
clrTypename(clrobj)

clrTypename(clrobj)
}
\arguments{
\item{clrobj}{CLR object}
}
\value{
type name

type name
}
\description{
Gets the CLR type name of an object, given an S4 clrobj object

Gets the CLR type name of an object, given an S4 clrobj object
}
\examples{
\dontrun{
library(rSharp)
testClassName <- "ClrFacade.TestObject"
testObj <- clrNew(testClassName)
clrTypename(testObj)
}
\dontrun{
library(rClr)
testClassName <- "Rclr.TestObject"
testObj <- clrNew(testClassName)
clrTypename(testObj)
}
}
