% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net-object.R
\name{NetObject}
\alias{NetObject}
\title{#' @title NetObject
#' @docType class
#'
#' @description
#' Base wrapper class for the S4 class `cobjRef` which holds pointers to .NET objects
#'
#' @importFrom R6
#'
#' @examples
#' myPrintable <- R6::R6Class(
#'   "myPrintable",
#'   inherit = Printable,
#'   public = list(
#'     x = NULL,
#'     y = NULL,
#'     print = function() {
#'       private$printClass()
#'       private$printLine("x", self$x)
#'       private$printLine("y", self$y)
#'       invisible(self)
#'     }
#'   )
#' )
#'
#' x <- myPrintable$new()
#' x}
\description{
#' @title NetObject
#' @docType class
#'
#' @description
#' Base wrapper class for the S4 class `cobjRef` which holds pointers to .NET objects
#'
#' @importFrom R6
#'
#' @examples
#' myPrintable <- R6::R6Class(
#'   "myPrintable",
#'   inherit = Printable,
#'   public = list(
#'     x = NULL,
#'     y = NULL,
#'     print = function() {
#'       private$printClass()
#'       private$printLine("x", self$x)
#'       private$printLine("y", self$y)
#'       invisible(self)
#'     }
#'   )
#' )
#'
#' x <- myPrintable$new()
#' x

#' @title NetObject
#' @docType class
#'
#' @description
#' Base wrapper class for the S4 class `cobjRef` which holds pointers to .NET objects
#'
#' @importFrom R6
#'
#' @examples
#' myPrintable <- R6::R6Class(
#'   "myPrintable",
#'   inherit = Printable,
#'   public = list(
#'     x = NULL,
#'     y = NULL,
#'     print = function() {
#'       private$printClass()
#'       private$printLine("x", self$x)
#'       private$printLine("y", self$y)
#'       invisible(self)
#'     }
#'   )
#' )
#'
#' x <- myPrintable$new()
#' x
}
\examples{

## ------------------------------------------------
## Method `NetObject$new`
## ------------------------------------------------

  testClassName <- "ClrFacade.Tests.RefClasses.LevelOneClass"
  o <- .External("r_create_clr_object", testClassName, PACKAGE = getRSharpSetting("nativePkgName"))
  x <- NetObject$new(o)
  print(x)
List the instance members of a .NET object


## ------------------------------------------------
## Method `NetObject$reflect`
## ------------------------------------------------

testClassName <- "ClrFacade.TestObject"
testObj <- newObjectFromName(testClassName)
clrReflect(testObj)
Print
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{type}}{The type of the .NET object. Read-only}

\item{\code{pointer}}{The external pointer to the .NET object. Read-only}

\item{\code{cobjRef}}{The underlying `cobjRef` object. Read-only
Initialize}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NetObject-new}{\code{NetObject$new()}}
\item \href{#method-NetObject-reflect}{\code{NetObject$reflect()}}
\item \href{#method-NetObject-print}{\code{NetObject$print()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetObject-new"></a>}}
\if{latex}{\out{\hypertarget{method-NetObject-new}{}}}
\subsection{Method \code{new()}}{
Initializes the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetObject$new(pointer)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pointer}}{The external pointer to the .NET object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The initialized object
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{  testClassName <- "ClrFacade.Tests.RefClasses.LevelOneClass"
  o <- .External("r_create_clr_object", testClassName, PACKAGE = getRSharpSetting("nativePkgName"))
  x <- NetObject$new(o)
  print(x)
List the instance members of a .NET object

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetObject-reflect"></a>}}
\if{latex}{\out{\hypertarget{method-NetObject-reflect}{}}}
\subsection{Method \code{reflect()}}{
List the instance members of a .NET object, i.e. its methods, fields and properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetObject$reflect()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of methods, fields, and properties of the object
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{testClassName <- "ClrFacade.TestObject"
testObj <- newObjectFromName(testClassName)
clrReflect(testObj)
Print
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetObject-print"></a>}}
\if{latex}{\out{\hypertarget{method-NetObject-print}{}}}
\subsection{Method \code{print()}}{
print prints a summary of the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetObject$print()}\if{html}{\out{</div>}}
}

}
}
