% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-net-object.R
\name{newObjectFromName}
\alias{newObjectFromName}
\title{Create a new NetObject R6 object given the type name.}
\usage{
newObjectFromName(typename, ..., R6objectClass = NetObject)
}
\arguments{
\item{typename}{type name, possibly namespace and assembly qualified type name, e.g. 'My.Namespace.MyClass,MyAssemblyName'.}

\item{...}{additional method arguments passed to the object constructor via the call to .External}

\item{R6objectClass}{the R6 class of the object to be created, defaults to `NetObject`.
Can be used to create custom R6 classes that inherit from `NetObject`.}
}
\value{
a `NetObject` R6 object
}
\description{
Create a new NetObject R6 object given the type name.
}
\examples{
testClassName <- getRSharpSetting("testObjectTypeName")
testObj <- newObjectFromName(testClassName)
# object with a constructor that has parameters
testObj <- newObjectFromName(testClassName, as.integer(123))
}
