% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSharp-env.R
\docType{data}
\name{rSharpSettingNames}
\alias{rSharpSettingNames}
\title{Names of the settings stored in rSharpEnv Can be used with `getRSharpSetting()`}
\format{
An object of class \code{character} of length 9.
}
\usage{
rSharpSettingNames
}
\description{
Names of the settings stored in rSharpEnv Can be used with `getRSharpSetting()`
}
\keyword{datasets}
