% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rClr-exported.R, R/rSharp-exported.R
\name{setClrRefClass}
\alias{setClrRefClass}
\title{Create reference classes for an object hierarchy}
\usage{
setClrRefClass(typeName, env = topenv(parent.frame()))

setClrRefClass(typeName, env = topenv(parent.frame()))
}
\arguments{
\item{typeName}{a CLR type name, recognizable by getType}

\item{env}{environment where the new generator is created.}
}
\value{
the object generator function

the object generator function
}
\description{
EXPERIMENTAL Create reference classes for an object hierarchy. Gratefully acknowledge Peter D. and its rJavax work.

EXPERIMENTAL Create reference classes for an object hierarchy. Gratefully acknowledge Peter D. and its rJavax work.
}
