% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-static.R
\name{setStatic}
\alias{setStatic}
\title{Sets the value of a field or property of an object or class}
\usage{
setStatic(type, name, value)
}
\arguments{
\item{type}{Type name, possibly namespace and assembly qualified type name, e.g. 'My.Namespace.MyClass,MyAssemblyName'.}

\item{name}{the name of a field/property of the object}

\item{value}{The value to set the field with}
}
\description{
Sets the value of a field or property of an object or class
}
\examples{
testClassName <- rSharpEnv$testObjectTypeName
setStatic(testClassName, "StaticPropertyIntegerOne", as.integer(42))
}
